pyproject.toml
[project]
name = "herbabot"
version = "0.1.0"
description = "Herbarium Telegram Bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.0",
    "piexif>=1.1.3",
    "pillow>=11.2.1",
    "pillow-heif>=0.22.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.1",
    "requests>=2.32.4",
]
